Combined the two jwtservices
